Move entity indexes into entites package
