+++ lenskit-core/src/main/java/org/lenskit/data/entities/GenericEntityIndex.java
@@ -18,24 +18,26 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.lenskit.data.dao;
+package org.lenskit.data.entities;
 
-import org.lenskit.data.entities.Entity;
+import com.google.common.collect.ImmutableListMultimap;
 
 import javax.annotation.Nonnull;
 import java.util.List;
 
 /**
- * An index to look up entities by attribute value.
- *
- * @see EntityIndexBuilder
+ * Generic implementation of the entity index.
  */
-public interface EntityIndex {
-    /**
-     * Get the entities with the associated attribute value.
-     * @param value The attribute value.
-     * @return The list of entities.
-     */
+class GenericEntityIndex implements EntityIndex {
+    private ImmutableListMultimap<Object,Entity> entities;
+
+    GenericEntityIndex(ImmutableListMultimap<Object,Entity> data) {
+        entities = data;
+    }
+
     @Nonnull
-    List<Entity> getEntities(@Nonnull Object value);
+    @Override
+    public List<Entity> getEntities(@Nonnull Object value) {
+        return entities.get(value);
+    }
 }
