+++ lenskit-core/src/main/java/org/lenskit/data/entities/LongEntityIndexBuilder.java
@@ -18,14 +18,12 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.lenskit.data.dao;
+package org.lenskit.data.entities;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.longs.Long2ObjectMap;
 import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
-import org.lenskit.data.entities.Entity;
-import org.lenskit.data.entities.TypedName;
 import org.lenskit.util.keys.SortedKeyIndex;
 
 /**
