+++ lenskit-core/src/test/java/org/lenskit/data/entities/GenericEntityIndexTest.java
@@ -18,21 +18,23 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.lenskit.data.dao;
+package org.lenskit.data.entities;
 
 import org.junit.Test;
 import org.lenskit.data.entities.Entity;
 import org.lenskit.data.entities.EntityFactory;
+import org.lenskit.data.entities.EntityIndex;
+import org.lenskit.data.entities.GenericEntityIndexBuilder;
 
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertThat;
 
-public class LongEntityIndexTest {
+public class GenericEntityIndexTest {
     private EntityFactory factory = new EntityFactory();
 
     @Test
     public void testEmpty() {
-        LongEntityIndexBuilder bld = new LongEntityIndexBuilder("user");
+        GenericEntityIndexBuilder bld = new GenericEntityIndexBuilder("user");
         EntityIndex index = bld.build();
         assertThat(index, notNullValue());
         assertThat(index.getEntities(20L), hasSize(0));
@@ -40,7 +42,7 @@ public class LongEntityIndexTest {
 
     @Test
     public void testAddEntity() {
-        LongEntityIndexBuilder bld = new LongEntityIndexBuilder("user");
+        GenericEntityIndexBuilder bld = new GenericEntityIndexBuilder("user");
         Entity rating = factory.rating(10, 100, 3.5);
         bld.add(rating);
         EntityIndex index = bld.build();
@@ -50,18 +52,8 @@ public class LongEntityIndexTest {
     }
 
     @Test
-    public void testNonLongValueEntity() {
-        LongEntityIndexBuilder bld = new LongEntityIndexBuilder("user");
-        Entity rating = factory.rating(10, 100, 3.5);
-        bld.add(rating);
-        EntityIndex index = bld.build();
-        assertThat(index, notNullValue());
-        assertThat(index.getEntities("10"), hasSize(0));
-    }
-
-    @Test
     public void testAddEntities() {
-        LongEntityIndexBuilder bld = new LongEntityIndexBuilder("user");
+        GenericEntityIndexBuilder bld = new GenericEntityIndexBuilder("user");
         Entity r1 = factory.rating(10, 100, 3.5);
         Entity r2 = factory.rating(10, 50, 4.5);
         Entity r3 = factory.rating(15, 100, 2.5);
